requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.1.0
langchain-groq==0.0.1
python-multipart==0.0.6
jinja2==3.1.2
python-dotenv==1.0.0
reportlab==4.0.7
weasyprint==60.2
pydantic==2.5.0
httpx==0.25.2
aiofiles==23.2.1
motor==3.3.1
pymongo==4.5.0
email-validator==2.0.0
